+++ client/Packages/com.beamable/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- `SimNetworkEventStream` has a `ISimFaultHandler` parameter that exposes error handling and callbacks for network outages. 
+
 ### Fixed
 - Fixed Content Manager exception when search doesn't match.